Updated code with latest changes

+++ src/main/java/com/BookMyShow/repositories/ShowRepository.java
@@ -2,8 +2,14 @@ package com.BookMyShow.repositories;
 
 import com.BookMyShow.models.Show;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ShowRepository extends JpaRepository<Show,Integer> {
+public interface ShowRepository extends JpaRepository<Show,Long> {
+
+
 }
