+++ src/main/java/com/BookMyShow/repositories/TheaterSeatRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TheaterSeatRepository extends JpaRepository<TheaterSeat,Integer> {
+public interface TheaterSeatRepository extends JpaRepository<TheaterSeat,Long> {
 
 
 }
